+++ lib/dial/middleware.rb
@@ -87,13 +87,13 @@ module Dial
     def clear_query_logs!
       [].tap do |query_logs|
         entry = section = count = nil
-        PROSOPITE_LOG_IO.string.lines.each do |line|
+        ProsopiteLogger.log_io.string.lines.each do |line|
           entry, section, count = process_query_log_line line, entry, section, count
           query_logs << entry if entry && section.nil?
         end
 
-        PROSOPITE_LOG_IO.truncate 0
-        PROSOPITE_LOG_IO.rewind
+        ProsopiteLogger.log_io.truncate 0
+        ProsopiteLogger.log_io.rewind
       end
     end