+++ lib/dial/railtie.rb
@@ -26,7 +26,7 @@ module Dial
         if ::ActiveRecord::Base.configurations.configurations.any? { |config| config.adapter == "postgresql" }
           require "pg_query"
         end
-        ::Prosopite.custom_logger = ProsopiteLogger.new PROSOPITE_LOG_IO
+        ::Prosopite.custom_logger = ProsopiteLogger.new
 
         # finalize configuration
         Dial._configuration.freeze